+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## [Unreleased]
+
+## [2.6.0] - 2021-11-23
 ### Added
 * Made component contracts (`Kreait\Firebase\Contract\*`) available via dependency injection.
 
@@ -54,7 +56,8 @@
 * Removed Support for `kreait/firebase-php:<5.0`
 * Removed project aliases
 
-[Unreleased]: https://github.com/kreait/firebase-bundle/compare/2.5.0...HEAD
+[Unreleased]: https://github.com/kreait/firebase-bundle/compare/2.6.0...HEAD
+[2.6.0]: https://github.com/kreait/firebase-bundle/compare/2.5.0...2.6.0
 [2.5.0]: https://github.com/kreait/firebase-bundle/compare/2.4.0...2.5.0
 [2.4.0]: https://github.com/kreait/firebase-bundle/compare/2.3.0...2.4.0
 [2.3.0]: https://github.com/kreait/firebase-bundle/compare/2.2.0...2.3.0